refactored AutoKeras adapter
- renamed entry file to be more telling
- refactored code from entry file into an AdapterUtils file
- refactored the adapter functionality itself to be more streamlined, every shared functionality will now be within the Abstract adapter and AdapterUtils file / class
- added launch.json for controller and autokeras to easily debug within VS Code
