+++ adapters/.dockerignore
@@ -1,0 +1,2 @@
+AutoKeras/.venv
+AutoKeras/.vscode
