+++ adapters/AutoKeras/.vscode/launch.json
@@ -1,0 +1,21 @@
+{
+    // Verwendet IntelliSense zum Ermitteln möglicher Attribute.
+    // Zeigen Sie auf vorhandene Attribute, um die zugehörigen Beschreibungen anzuzeigen.
+    // Weitere Informationen finden Sie unter https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "AutoKeras Adapter",
+            "type": "python",
+            "request": "launch",
+            "program": "AutoKerasServer.py",
+            "console": "integratedTerminal",
+            "justMyCode": true,
+            "env": {
+                "PYTHONPATH": "AutoMLs;Utils;../Utils/Utils;../Utils/AutoMLs;../GRPC/Adapter;",
+                "GRPC_SERVER_PORT": "50052",
+                "PYTHON_ENV": ".venv/Scripts/python"
+            },
+        }
+    ]
+}
