+++ adapters/AutoKeras/Dockerfile
@@ -18,4 +18,4 @@ ENV RUNTIME=DOCKER
 ENV PYTHONUNBUFFERED=1
 ENV PYTHONPATH "AutoMLs:Utils"
 ENV PYTHON_ENV "python"
-ENTRYPOINT ["python", "Adapter_AutoKeras.py"]
+ENTRYPOINT ["python", "AutoKerasServer.py"]
