+++ adapters/AutoKeras/config/config.json
@@ -2,7 +2,7 @@
   "adapter-name": "autokeras",
   "job-file-name": "keras-job.json",
   "env-varname-cluster": "RUNTIME",
-  "grpc-server-address": "0.0.0.0",
+  "grpc-server-address": "0.0.0.0",     
   "export-zip-file-name": "keras-export",
   "output-path": "app-data/output",
   "job-file-path": "app-data/jobs",
