+++ frontend/src/Server/BlazorBoilerplate.Server/Properties/launchSettings.json
@@ -3,7 +3,7 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:54474",
+      "applicationUrl": "http://localhost:54475",
       "sslPort": 0
     }
   },
